Medium changes
Added CustomThread, a custom Thread class with the boolean "running" and start() and stopThread() methods. Started testing out 0.6s ticks with GameUpdater. Changed all classes that extend Thread to extend CustomThread. Added a list of all CustomThreads to the server, so it can stop all threads when stopServer() is called.